requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-magic>=0.4.27
python-multipart>=0.0.6
google-generativeai>=0.3.0
